+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Add `url` field to Mastodon API Status object.
+
 ### Changed
 
 - Write instance URL to log at startup.