+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - Write instance URL to log at startup.
+- Change MSRV to 1.60.0.
 
 ## [2.4.0] - 2023-11-16